+++ CHANGELOG.md
@@ -48,6 +48,7 @@
 * Fix user search pagination and total count display beyond 1000 records. Refs UIU-3387.
 * Hide permissions and roles accordions on user create page. Refs UIU-3449.
 * Add support for use circulation-bff for mark-as-missing.  Refs UIU-3444.
+* In the list of a user's open loans, the "Use at location" includes service-point and (when relevant) exipiry date. Fixes UIU-3418.
 
 ## [12.1.11] (https://github.com/folio-org/ui-users/tree/v12.1.11) (2025-08-27)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v12.1.10...v12.1.11)